Add check format controllers and actions names. Rewrite error handler. Apdated documentatшщт for framework

+++ controllers/NewsController.php
@@ -8,11 +8,17 @@ class NewsController
 {
 	//Метод устанавливает что новость прочитана
 	public static function actionSetRead() {
-		if((isset($_POST['newsId'])) && ($_POST['newsId'] !== '')) {
+		$dataType = gettype($_POST['newsId']);
+		if((isset($_POST['newsId'])) && ($_POST['newsId'] !== '') && ($dataType === 'string')) {
 			$cleanNewsId = Maker::$app -> filter($_POST['newsId']);
 			Maker::$app -> query('UPDATE news SET news.read = ? WHERE news_id = ? LIMIT 1', [1, (int)$cleanNewsId]);
 		} else {
-			trigger_error('Не установлен или пустой идентификатор новости||0');
+			if($dataType !== 'string') {
+                Maker::$app -> error('В методе '.__METHOD__.' тип id новости не является string');
+            }
+            if(($_POST['newsId'] === '') || (isset($_POST['newsId']) === false)) {
+                Maker::$app -> error('В методе '.__METHOD__.' не определен id новости');
+            }
 		}
 	}
 
@@ -23,7 +29,8 @@ class NewsController
 
 	//Метод переноса новости в архив
 	public static function actionSendToArchive() {
-		if((isset($_POST['newsId'])) && ($_POST['newsId'] !== '')) {
+		$dataType = gettype($_POST['newsId']);
+		if((isset($_POST['newsId'])) && ($_POST['newsId'] !== '') && ($dataType === 'string')) {
 			$cleanNewsId = Maker::$app -> filter($_POST['newsId']);
 			//Получаем данные новости, которая будет отправлено в архив
 			$newsData = Maker::$app -> query('SELECT news.news_title, news.news_description, news.news_link, news.publication_date, news.rss_url_list_id FROM news WHERE news.news_id = ?', [(int)$cleanNewsId]);
@@ -31,17 +38,28 @@ class NewsController
 			//Отправляем новость в архив
 			Maker::$app -> query('INSERT INTO news_archive VALUES (?, ?, ?, ?, ?, ?)', [NULL, (int)$newsData[0]['rss_url_list_id'], $newsData[0]['news_title'], $newsData[0]['news_description'], $newsData[0]['news_link'], $newsData[0]['publication_date']]);
 		} else {
-			trigger_error('Не установлен или пустой идентификатор новости||0');
+			if($dataType !== 'string') {
+                Maker::$app -> error('В методе '.__METHOD__.' тип id новости не является string');
+            }
+            if(($_POST['newsId'] === '') || (isset($_POST['newsId']) === false)) {
+                Maker::$app -> error('В методе '.__METHOD__.' не определен id новости');
+            }
 		}
 	}
 
 	//Метод удаления новости из архива
 	public static function actionDeleteFromArchive() {
-		if((isset($_POST['newsId'])) && ($_POST['newsId'] !== '')) {
+		$dataType = gettype($_POST['newsId']);
+		if((isset($_POST['newsId'])) && ($_POST['newsId'] !== '') && ($dataType === 'string')) {
 			$cleanNewsId = Maker::$app -> filter($_POST['newsId']);
 			Maker::$app -> query('DELETE FROM news_archive WHERE news_archive.news_archive_id = ?', [(int)$cleanNewsId]);
 		} else {
-			trigger_error('Не установлен или пустой идентификатор новости||0');
+			if($dataType !== 'string') {
+                Maker::$app -> error('В методе '.__METHOD__.' тип id новости не является string');
+            }
+            if(($_POST['newsId'] === '') || (isset($_POST['newsId']) === false)) {
+                Maker::$app -> error('В методе '.__METHOD__.' не определен id новости');
+            }
 		}
 	}
 }
